// web/src/lib/api/generated/models/poskasir-internal-user-repository-user-role.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const POSKasirInternalUserRepositoryUserRole = {
    UserRoleAdmin: 'admin',
    UserRoleCashier: 'cashier',
    UserRoleManager: 'manager'
} as const;

export type POSKasirInternalUserRepositoryUserRole = typeof POSKasirInternalUserRepositoryUserRole[keyof typeof POSKasirInternalUserRepositoryUserRole];